Try the local binder

+++ src/core/libutils/SharedBuffer.cpp
@@ -29,7 +29,7 @@ SharedBuffer* SharedBuffer::alloc(size_t size)
 {
     // Don't overflow if the combined size of the buffer / header is larger than
     // size_max.
-    LOG_ALWAYS_FATAL_IF((size >= (SIZE_MAX - sizeof(SharedBuffer))),
+    LOG_ALWAYS_FATAL_IF((size >= ((size_t)SIZE_MAX - (size_t)sizeof(SharedBuffer))),
                         "Invalid buffer size %zu", size);
 
     SharedBuffer* sb = static_cast<SharedBuffer *>(malloc(sizeof(SharedBuffer) + size));
